+++ source/EngineGpuKernels/SimulationData.cu
@@ -14,9 +14,8 @@ void SimulationData::init(int2 const& worldSize_, uint64_t timestep_)
     cellMap.init(worldSize);
     particleMap.init(worldSize);
 
-    CudaMemoryManager::getInstance().acquireMemory<ColorVector<float>>(1, externalEnergy);
-    CudaMemoryManager::getInstance().acquireMemory<double>(1, residualEnergy);
-    CHECK_FOR_CUDA_ERROR(cudaMemset(residualEnergy, 0, sizeof(double)));
+    CudaMemoryManager::getInstance().acquireMemory<double>(1, externalEnergy);
+    CHECK_FOR_CUDA_ERROR(cudaMemset(externalEnergy, 0, sizeof(double)));
  
     processMemory.init();
     numberGen1.init(40312357);   //some array size for random numbers (~ 40 MB)
@@ -42,9 +41,7 @@ __device__ void SimulationData::prepareForNextTimestep()
     for (int i = 0; i < CellFunction_WithoutNone_Count; ++i) {
         cellFunctionOperations[i].setMemory(processMemory.getTypedSubArray<CellFunctionOperation>(maxCellFunctionOperations), maxCellFunctionOperations);
     }
-    for (int i = 0; i < MAX_COLORS; ++i) {
-        (*externalEnergy)[i] = cudaSimulationParameters.cellFunctionConstructorExternalEnergy[i];
-    }
+    *externalEnergy = cudaSimulationParameters.externalEnergy;
 
     objects.saveNumEntries();
 }
@@ -113,7 +110,6 @@ void SimulationData::free()
     numberGen2.free();
     processMemory.free();
     CudaMemoryManager::getInstance().freeMemory(externalEnergy);
-    CudaMemoryManager::getInstance().freeMemory(residualEnergy);
 
     structuralOperations.free();
     for (int i = 0; i < CellFunction_WithoutNone_Count; ++i) {