+++ source/Gui/SimulationParametersWindow.cpp
@@ -247,13 +247,25 @@ void _SimulationParametersWindow::processBase(
                     .name("Cell coloring")
                     .textWidth(RightColumnWidth)
                     .defaultValue(origParameters.cellColoring)
-                    .values({"None", "Standard cell colors", "Mutants", "Cell state", "Genome complexity", "Highlight cell function"})
+                    .values(
+                        {"None",
+                         "Standard cell colors",
+                         "Mutants",
+                         "Mutants and cell functions",
+                         "Cell states",
+                         "Genome complexities",
+                         "Single cell function",
+                         "All cell functions"})
                     .tooltip("Here, one can set how the cells are to be colored during rendering. \n\n"
-                            ICON_FA_CHEVRON_RIGHT " Standard cell colors: Each cell is assigned one of 7 default colors, which is displayed with this option. \n\n" ICON_FA_CHEVRON_RIGHT
-                             " Mutants: Different mutants are represented by different colors (only larger structural mutations such as translations or duplications are taken into account).\n\n" ICON_FA_CHEVRON_RIGHT
-                        " Cell state: green = under construction, blue = ready, red = dying\n\n" ICON_FA_CHEVRON_RIGHT
-                        " Genome complexity: This property can be utilized by attacker cells when the parameter 'Complex genome protection' is "
-                        "activated (see tooltip there). The coloring is as follows: blue = creature with low bonus (usually small or simple genome structure), red = large bonus"),
+                        ICON_FA_CHEVRON_RIGHT " Standard cell colors: Each cell is assigned one of 7 default colors, which is displayed with this option. \n\n"
+                        ICON_FA_CHEVRON_RIGHT " Mutants: Different mutants are represented by different colors (only larger structural mutations such as translations or duplications are taken into account).\n\n"
+                        ICON_FA_CHEVRON_RIGHT " Mutants and cell functions: It combines the coloring for mutants and cell functions.\n\n"
+                        ICON_FA_CHEVRON_RIGHT " Cell states: green = under construction, blue = ready, red = dying\n\n"
+                        ICON_FA_CHEVRON_RIGHT " Genome complexities: This property can be utilized by attacker cells when the parameter 'Complex genome protection' is "
+                        "activated (see tooltip there). The coloring is as follows: blue = creature with low bonus (usually small or simple genome structure), red = large bonus\n\n"
+                        ICON_FA_CHEVRON_RIGHT " Single cell function: A specific type of cell function can be highlighted, which is selected in the next parameter.\n\n"
+                        ICON_FA_CHEVRON_RIGHT " All cell functions: The cells are colored according to their cell function.\n\n"
+                    ),
                 parameters.cellColoring);
             if (parameters.cellColoring == CellColoring_CellFunction) {
                 AlienImGui::Switcher(
@@ -919,17 +931,6 @@ void _SimulationParametersWindow::processBase(
         if (AlienImGui::BeginTreeNode(AlienImGui::TreeNodeParameters().text("Cell function: Constructor"))) {
             AlienImGui::SliderFloat(
                 AlienImGui::SliderFloatParameters()
-                    .name("Pump energy")
-                    .textWidth(RightColumnWidth)
-                    .colorDependence(true)
-                    .min(0.00f)
-                    .max(1.0f)
-                    .defaultValue(origParameters.cellFunctionConstructorPumpEnergyFactor)
-                    .tooltip("This parameter controls the energy pump system. It describes the fraction of the energy cost for a offspring which a constructor "
-                             "can get for free. This additional energy is obtain from radiation of other cells."),
-                parameters.cellFunctionConstructorPumpEnergyFactor);
-            AlienImGui::SliderFloat(
-                AlienImGui::SliderFloatParameters()
                     .name("Offspring distance")
                     .textWidth(RightColumnWidth)
                     .colorDependence(true)
@@ -1351,31 +1352,54 @@ void _SimulationParametersWindow::processBase(
                     AlienImGui::SliderFloatParameters()
                         .name("External energy amount")
                         .textWidth(RightColumnWidth)
-                        .colorDependence(true)
                         .min(0.0f)
                         .max(100000000.0f)
                         .format("%.0f")
                         .logarithmic(true)
                         .infinity(true)
-                        .defaultValue(origParameters.cellFunctionConstructorExternalEnergy)
+                        .defaultValue(&origParameters.externalEnergy)
                         .tooltip(
-                            "This parameter can be used to set the amount of energy (per color) of an external energy source. This type of energy is "
-                            "transferred to all constructor cells at a certain rate.\n\nTip: You can explicitly enter a numerical value by selecting the "
+                            "This parameter can be used to set the amount of energy of an external energy source. This type of energy can be "
+                            "transferred to all constructor cells at a certain rate (see inflow settings).\n\nTip: You can explicitly enter a numerical value by selecting the "
                             "slider and then pressing TAB.\n\nWarning: Too much external energy can result in a massive production of cells and slow down or "
                             "even crash the simulation."),
-                    parameters.cellFunctionConstructorExternalEnergy);
+                    &parameters.externalEnergy);
                 AlienImGui::SliderFloat(
                     AlienImGui::SliderFloatParameters()
-                        .name("External energy supply rate")
+                        .name("Inflow")
                         .textWidth(RightColumnWidth)
                         .colorDependence(true)
                         .min(0.0f)
                         .max(1.0f)
-                        .defaultValue(origParameters.cellFunctionConstructorExternalEnergySupplyRate)
+                        .defaultValue(origParameters.externalEnergyInflowFactor)
                         .tooltip(
-                            "The energy from the external source is transferred to all constructor cells at a rate defined here: 0 = no energy transfer, 1 = "
-                            "constructor cells receive all the required energy"),
-                    parameters.cellFunctionConstructorExternalEnergySupplyRate);
+                            "Here one can specify the fraction of energy transferred to constructor cells.\n\nFor example, a value of 0.05 means that each time "
+                            "a constructor cell tries to build a new cell, 5% of the required energy is transferred for free from the external energy source."),
+                    parameters.externalEnergyInflowFactor);
+                AlienImGui::SliderFloat(
+                    AlienImGui::SliderFloatParameters()
+                        .name("Conditional inflow")
+                        .textWidth(RightColumnWidth)
+                        .colorDependence(true)
+                        .min(0.00f)
+                        .max(1.0f)
+                        .defaultValue(origParameters.externalEnergyConditionalInflowFactor)
+                        .tooltip("Here one can specify the fraction of energy transferred to constructor cells if they can provide the remaining energy for the "
+                                 "construction process.\n\nFor example, a value of 0.6 means that a constructor cell receives 60% of the energy required to "
+                                 "build the new cell for free from the external energy source. However, it must provide 40% of the energy required by itself. "
+                                 "Otherwise, no energy will be transferred."),
+                    parameters.externalEnergyConditionalInflowFactor);
+                AlienImGui::SliderFloat(
+                    AlienImGui::SliderFloatParameters()
+                        .name("Backflow")
+                        .textWidth(RightColumnWidth)
+                        .colorDependence(true)
+                        .min(0.0f)
+                        .max(1.0f)
+                        .defaultValue(origParameters.externalEnergyBackflowFactor)
+                        .tooltip("The proportion of energy that flows back to the external energy source when a cell loses energy or dies. The remaining "
+                                 "fraction of the energy is used to create a new energy particle."),
+                    parameters.externalEnergyBackflowFactor);
                 AlienImGui::EndTreeNode();
             }
         }
@@ -2099,7 +2123,8 @@ void _SimulationParametersWindow::processAddonList(
                     .name("External energy control")
                     .textWidth(0)
                     .defaultValue(origParameters.features.externalEnergyControl)
-                    .tooltip("This addon is used to add an external energy source. The energy is gradually transferred to the cells in the simulation."),
+                    .tooltip("This addon is used to add an external energy source. Its energy can be gradually transferred to the constructor cells in the "
+                             "simulation. Vice versa, the energy from radiation and dying cells can also be transferred back to the external source."),
                 parameters.features.externalEnergyControl);
             AlienImGui::Checkbox(
                 AlienImGui::CheckboxParameters()
@@ -2164,13 +2189,12 @@ void _SimulationParametersWindow::validationAndCorrection(SimulationParameters& 
         parameters.baseValues.radiationAbsorption[i] = std::max(0.0f, std::min(1.0f, parameters.baseValues.radiationAbsorption[i]));
         parameters.radiationAbsorptionHighVelocityPenalty[i] = std::max(0.0f, parameters.radiationAbsorptionHighVelocityPenalty[i]);
         parameters.radiationAbsorptionLowConnectionPenalty[i] = std::max(0.0f, parameters.radiationAbsorptionLowConnectionPenalty[i]);
-        parameters.cellFunctionConstructorPumpEnergyFactor[i] = std::max(0.0f, std::min(1.0f, parameters.cellFunctionConstructorPumpEnergyFactor[i]));
+        parameters.externalEnergyConditionalInflowFactor[i] = std::max(0.0f, std::min(1.0f, parameters.externalEnergyConditionalInflowFactor[i]));
         parameters.cellFunctionAttackerSensorDetectionFactor[i] = std::max(0.0f, std::min(1.0f, parameters.cellFunctionAttackerSensorDetectionFactor[i]));
         parameters.cellFunctionDetonatorChainExplosionProbability[i] =
             std::max(0.0f, std::min(1.0f, parameters.cellFunctionDetonatorChainExplosionProbability[i]));
-        parameters.cellFunctionConstructorExternalEnergy[i] = std::max(0.0f, parameters.cellFunctionConstructorExternalEnergy[i]);
-        parameters.cellFunctionConstructorExternalEnergySupplyRate[i] =
-            std::max(0.0f, std::min(1.0f, parameters.cellFunctionConstructorExternalEnergySupplyRate[i]));
+        parameters.externalEnergyInflowFactor[i] =
+            std::max(0.0f, std::min(1.0f, parameters.externalEnergyInflowFactor[i]));
         parameters.baseValues.cellMinEnergy[i] = std::min(parameters.baseValues.cellMinEnergy[i], parameters.cellNormalEnergy[i] * 0.95f);
         parameters.particleSplitEnergy[i] = std::max(0.0f, parameters.particleSplitEnergy[i]);
         parameters.baseValues.radiationAbsorptionLowGenomeComplexityPenalty[i] =
@@ -2178,6 +2202,7 @@ void _SimulationParametersWindow::validationAndCorrection(SimulationParameters& 
         parameters.baseValues.radiationAbsorptionLowVelocityPenalty[i] =
             std::max(0.0f, std::min(1.0f, parameters.baseValues.radiationAbsorptionLowVelocityPenalty[i]));
     }
+    parameters.externalEnergy = std::max(0.0f, parameters.externalEnergy);
     parameters.baseValues.cellMaxBindingEnergy = std::max(10.0f, parameters.baseValues.cellMaxBindingEnergy);
     parameters.timestepSize = std::max(0.0f, parameters.timestepSize);
     parameters.cellMaxAgeBalancerInterval = std::max(1000, std::min(1000000, parameters.cellMaxAgeBalancerInterval));